feat: add feature for loading journals in each collection

+++ server/src/main/java/io/kioke/feature/collection/service/CollectionService.java
@@ -2,10 +2,12 @@ package io.kioke.feature.collection.service;
 
 import io.kioke.exception.collection.CollectionNotFoundException;
 import io.kioke.feature.collection.domain.Collection;
+import io.kioke.feature.collection.domain.CollectionEntry;
 import io.kioke.feature.collection.dto.CollectionDto;
 import io.kioke.feature.collection.dto.CreateCollectionRequest;
 import io.kioke.feature.collection.repository.CollectionRepository;
 import io.kioke.feature.collection.util.CollectionMapper;
+import io.kioke.feature.journal.domain.Journal;
 import io.kioke.feature.user.domain.User;
 import io.kioke.feature.user.dto.UserPrincipal;
 import java.util.List;
@@ -37,29 +39,37 @@ public class CollectionService {
 
   @Transactional
   public CollectionDto createCollection(UserPrincipal requester, CreateCollectionRequest request) {
+    User user = User.getUserReference(requester.userId());
+
+    boolean isFirstCollection = collectionRepository.findDefaultCollectionByUser(user).isEmpty();
+
     Collection collection =
-        Collection.builder()
-            .user(User.getUserReference(requester.userId()))
-            .name(request.name())
-            .build();
+        Collection.builder().user(user).name(request.name()).isDefault(isFirstCollection).build();
     collection = collectionRepository.save(collection);
 
     return collectionMapper.toDto(collection);
   }
 
   @Transactional
-  public void deleteCollection(String collectionId) {
+  public void deleteCollection(UserPrincipal requester, String collectionId) {
+    User user = User.getUserReference(requester.userId());
+
+    if (collectionRepository.countByUser(user) == 1) {
+      throw new IllegalStateException("User has no collections to delete.");
+    }
+
     collectionRepository.deleteById(collectionId);
   }
 
-  // @Transactional
-  // public void addEntryToCollection(String userId, String journalId, String collectionId)
-  //     throws CollectionNotFoundException {
-  //   Collection collection =
-  //       collectionRepository
-  //           .findByUserAndId(userId, collectionId)
-  //           .orElseThrow(() -> new CollectionNotFoundException());
-  //   collection.addJournal(journal);
-  //   collectionRepository.save(collection);
-  // }
+  @Transactional
+  public void addJournalToCollection(UserPrincipal requester, Journal journal, String collectionId)
+      throws CollectionNotFoundException {
+    Collection collection = collectionRepository.findById(collectionId).orElseThrow();
+
+    CollectionEntry collectionEntry =
+        CollectionEntry.builder().collection(collection).journal(journal).build();
+    collection.getJournals().add(collectionEntry);
+
+    collectionRepository.save(collection);
+  }
 }
